feat: hook -> util 함수로 수정
