+++ src/components/MakeClub/Modal/index.tsx
@@ -23,7 +23,7 @@ import {
   DialogTrigger,
 } from '@/components/ui/dialog';
 
-import { IsDateBeforeToday } from '@/lib/utils';
+import { isDateBeforeToday } from '@/lib/utils';
 
 import { usePostGatherings } from '@/hooks/useGatherings';
 
@@ -69,7 +69,7 @@ export default function MakeClubModal({ trigger }: Props) {
   const flattenedCategories = flattenCategories(CATEGORY);
 
   // 달력 에러 메세지
-  const isSelectedDateBeforeToday = IsDateBeforeToday({ date });
+  const isSelectedDateBeforeToday = isDateBeforeToday({ date });
   let dateErrorMsg = null;
   if (!date) {
     dateErrorMsg = ERROR_MESSAGE.date.required;
